Add check scripts for existence

+++ easy_wireguard.sh
@@ -6,14 +6,23 @@ RED='\033[0;31m'
 NC='\033[0m'
 
 echo -e "${GREEN}Downloading scripts...${NC}"
-mkdir easy_wireguard
+
+if ! [ -d "easy_wireguard" ]; then
+    mkdir easy_wireguard
+fi
 cd easy_wireguard
-curl -O https://raw.githubusercontent.com/bllizard22/easy-wireguard-server/main/setup_server.sh
-chmod +x setup_server.sh
-curl -O https://raw.githubusercontent.com/bllizard22/easy-wireguard-server/main/add_client.sh
-chmod +x add_client.sh
-curl -O https://raw.githubusercontent.com/bllizard22/easy-wireguard-server/main/remove_server.sh
-chmod +x remove_server.sh
+if ! [ -f "setup_server.sh" ]; then
+    curl -O https://raw.githubusercontent.com/bllizard22/easy-wireguard-server/main/setup_server.sh
+	chmod +x setup_server.sh
+fi
+if ! [ -f "add_client.sh" ]; then
+    curl -O https://raw.githubusercontent.com/bllizard22/easy-wireguard-server/main/add_client.sh
+	chmod +x add_client.sh
+fi
+if ! [ -f "remove_server.sh" ]; then
+    curl -O https://raw.githubusercontent.com/bllizard22/easy-wireguard-server/main/remove_server.sh
+	chmod +x remove_server.sh
+fi
 cd ..
 
 echo -en "${GREEN}Choose the action:
@@ -29,7 +38,8 @@ if [ $OPTION == "1" ]; then
 elif [ $OPTION == "2" ]; then
 	./easy_wireguard/add_client.sh
 elif [ $OPTION == "3" ]; then
-	./easy_wireguard/remove_server.sh	
+	./easy_wireguard/remove_server.sh
+	rm -r ./easy_wireguard
 else
 	echo -e "${PURPLE}Exit without actions${NC}"
 fi
